pyproject.toml
[build-system]
requires = ["setuptools", "wheel", "poetry-core", "poetry-plugin-export"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
name = "tournaments-backend"
version = "0.3.0"
description = "Tournaments Backend"
readme = "README.md"
authors = [
    "Luca Cireddu <sardylan@gmail.com>",
    "Gavino Isoni <sardylan@gmail.com>",
    "Marco Di Francesco <sardylan@gmail.com>",
    "Matteo Farci <sardylan@gmail.com>",
]
classifiers = []
keywords = []
include = []

[tool.poetry.dependencies]
python = ">=3.12,<=3.13"

asn1crypto = "==1.5.1"
Babel = "==2.10.3"
cbor2 = "==5.6.2"
chardet = "==5.2.0"
cryptography = "==42.0.8"
decorator = "==5.1.1"
docutils = "==0.20.1"
freezegun = "==1.2.1"
geoip2 = "==2.9.0"
gevent = "==24.2.1"
greenlet = "==3.0.3"
idna = "==3.6"
inotify = "==0.2.10"
Jinja2 = "==3.1.2"
libsass = "==0.22.0"
lxml = "==5.2.1"
lxml-html-clean = "==0.4.1"
MarkupSafe = "==2.1.5"
num2words = "==0.5.13"
ofxparse = "==0.21"
openpyxl = "==3.1.2"
passlib = "==1.7.4"
Pillow = "==10.2.0"
polib = "==1.2.0"
psutil = "==5.9.8"
psycopg2 = "==2.9.9"
pyopenssl = "==24.1.0"
PyPDF2 = "==2.12.1"
pyserial = "==3.5"
python-dateutil = "==2.8.2"
python-stdnum = "==1.19"
pytz = "==2025.1"
pyusb = "==1.2.1"
qrcode = "==7.4.2"
reportlab = "==4.1.0"
requests = "==2.31.0"
rjsmin = "==1.2.0"
urllib3 = "==2.0.7"
vobject = "==0.9.6.1"
Werkzeug = "==3.0.1"
xlrd = "==2.0.1"
XlsxWriter = "==3.1.9"
xlwt = "==1.3.0"
zeep = "==4.2.1"

fastapi = "==0.115.8"
pydantic = "==2.11.0a1"
pyjwt = "==2.10.1"
pyyaml = "==6.0.2"
uvicorn = "==0.34.0"
uvloop = "==0.21.0"
python-magic = "==0.4.27"

[tool.poetry.group.dev.dependencies]
pre-commit = "==4.1.0"
pytest = "==8.3.4"
pytest-asyncio = "==0.25.3"
httpx = "==0.28.1"

[tool.pytest.ini_options]
asyncio_mode = "auto"
